Day 09.

+++ AdventOfCode/AdventOfCode/Solvers/Day02/Day02Solver.cs
@@ -7,24 +7,24 @@ using AdventOfCode.Computer;
 namespace AdventOfCode.Solvers {
   public class Day02Solver : Solver {
     public string SolvePartOne(string[] input) {
-      int[] program = string.Join(",", input).Split(",")
-        .Select(x => Int32.Parse(x)).ToArray();
+      long[] program = string.Join(",", input).Split(",")
+        .Select(x => Int64.Parse(x)).ToArray();
 
       CompileWithNounAndVerb(program, 12, 2);
       return program[0].ToString();
     }
 
     public string SolvePartTwo(string[] input) {
-      int[] prog = string.Join(",", input).Split(",")
-        .Select(x => Int32.Parse(x)).ToArray();
+      long[] prog = string.Join(",", input).Split(",")
+        .Select(x => Int64.Parse(x)).ToArray();
 
-      int desired = 19690720;
-      int noun = 0;
-      int verb = 0;
-      int[] compiled = (int[]) prog.Clone();
+      long desired = 19690720;
+      long noun = 0;
+      long verb = 0;
+      long[] compiled = (long[]) prog.Clone();
       for(noun = 0; desired != compiled[0] && noun <= 99; noun++) {
         for(verb = 0; desired != compiled[0] && verb <= 99; verb++) {
-          compiled = (int[]) prog.Clone();
+          compiled = (long[]) prog.Clone();
           CompileWithNounAndVerb(compiled, noun, verb);
         } 
       }
@@ -32,7 +32,7 @@ namespace AdventOfCode.Solvers {
       return (100 * --noun + --verb).ToString();
     }
 
-    private void CompileWithNounAndVerb(int[] program, int noun, int verb) {
+    private void CompileWithNounAndVerb(long[] program, long noun, long verb) {
       program[1] = noun;
       program[2] = verb;
